+++ Virtuoso/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Upcoming
+
+### Changed
+
+- Connects sync on first toot instead of via polling
+
+### Removed
+
+- ConnectInterval config entry (no longer used)
+
 ## [0.1.1] - 2025-08-04
 
 ### Changed